[impl] Cluster Bootstrap (#29)
* impl(ws-raft-raft-protocol-phase-project-scaffolding-and-core-types-stage-cargo-workspace-and-root-configuration): Cargo Workspace and Root Configuration (#2)


* impl(ws-raft-raft-protocol-phase-persistence-crash-recovery-and-bootstrap-xraft-core-stage-cluster-bootstrap): Cluster Bootstrap

* forge: nudge mergeable on ws/raft-RAFT-PROTOCOL/phase-persistence-crash-recovery-and-bootstrap-xraft-core-stage-cluster-bootstrap (origin/feature/raft-protocol already ancestor)

* forge: nudge mergeable on ws/raft-RAFT-PROTOCOL/phase-persistence-crash-recovery-and-bootstrap-xraft-core-stage-cluster-bootstrap (origin/feature/raft-protocol already ancestor)

* forge: nudge mergeable on ws/raft-RAFT-PROTOCOL/phase-persistence-crash-recovery-and-bootstrap-xraft-core-stage-cluster-bootstrap (origin/feature/raft-protocol already ancestor)

* forge: nudge mergeable on ws/raft-RAFT-PROTOCOL/phase-persistence-crash-recovery-and-bootstrap-xraft-core-stage-cluster-bootstrap (origin/feature/raft-protocol already ancestor)

* forge: nudge mergeable on ws/raft-RAFT-PROTOCOL/phase-persistence-crash-recovery-and-bootstrap-xraft-core-stage-cluster-bootstrap (origin/feature/raft-protocol already ancestor)

* forge: nudge mergeable on ws/raft-RAFT-PROTOCOL/phase-persistence-crash-recovery-and-bootstrap-xraft-core-stage-cluster-bootstrap (origin/feature/raft-protocol already ancestor)

* forge: nudge mergeable on ws/raft-RAFT-PROTOCOL/phase-persistence-crash-recovery-and-bootstrap-xraft-core-stage-cluster-bootstrap (origin/feature/raft-protocol already ancestor)

---------

+++ xraft-core/src/voter.rs
@@ -1,6 +1,5 @@
-use std::net::SocketAddr;
-
 use serde::{Deserialize, Serialize};
+use std::net::SocketAddr;
 
 use crate::types::NodeId;
 
@@ -11,9 +10,9 @@ pub struct VoterInfo {
     pub endpoint: SocketAddr,
 }
 
-/// A snapshot of the current voter set, committed as a control record.
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+/// Complete voter set record, committed via the log.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct VotersRecord {
-    pub version: u64,
+    pub version: u32,
     pub voters: Vec<VoterInfo>,
 }
